Finished saving & loading of buildings

+++ Assets/Scripts/SaveAbles/SaveGame.cs
@@ -9,7 +9,7 @@ public class SaveGame
     public int AmountStone { get; set; }
     public int AmountGold { get; set; }
     public int AmountFood { get; set; }
-    public int AmountHappiness { get; set; }
+    public float AmountHappiness { get; set; }
     public int Day { get; set; }
     public int Year { get; set; }
 }
